+++ go/v4/onetrading.go
@@ -427,7 +427,7 @@ func (this *Onetrading) Describe() any {
 					"marginMode":       false,
 					"limit":            100,
 					"daysBack":         100000,
-					"daysBackCanceled": 1 / 12,
+					"daysBackCanceled": float64(1) / 12,
 					"untilDays":        30,
 					"trigger":          false,
 					"trailing":         false,
