+++ go/v4/poloniex.go
@@ -343,13 +343,13 @@ func (this *Poloniex) Describe() any {
 			"swapPublic": map[string]any{
 				"get": map[string]any{
 					"v3/market/allInstruments": map[string]any{
-						"cost": 2 / 3,
+						"cost": float64(2) / 3,
 					},
 					"v3/market/instruments": map[string]any{
-						"cost": 2 / 3,
+						"cost": float64(2) / 3,
 					},
 					"v3/market/orderBook": map[string]any{
-						"cost": 2 / 3,
+						"cost": float64(2) / 3,
 					},
 					"v3/market/candles": map[string]any{
 						"cost": 10,
@@ -364,40 +364,40 @@ func (this *Poloniex) Describe() any {
 						"cost": 10,
 					},
 					"v3/market/trades": map[string]any{
-						"cost": 2 / 3,
+						"cost": float64(2) / 3,
 					},
 					"v3/market/liquidationOrder": map[string]any{
-						"cost": 2 / 3,
+						"cost": float64(2) / 3,
 					},
 					"v3/market/tickers": map[string]any{
-						"cost": 2 / 3,
+						"cost": float64(2) / 3,
 					},
 					"v3/market/markPrice": map[string]any{
-						"cost": 2 / 3,
+						"cost": float64(2) / 3,
 					},
 					"v3/market/indexPrice": map[string]any{
-						"cost": 2 / 3,
+						"cost": float64(2) / 3,
 					},
 					"v3/market/indexPriceComponents": map[string]any{
-						"cost": 2 / 3,
+						"cost": float64(2) / 3,
 					},
 					"v3/market/fundingRate": map[string]any{
-						"cost": 2 / 3,
+						"cost": float64(2) / 3,
 					},
 					"v3/market/fundingRate/history": map[string]any{
-						"cost": 2 / 3,
+						"cost": float64(2) / 3,
 					},
 					"v3/market/openInterest": map[string]any{
-						"cost": 2 / 3,
+						"cost": float64(2) / 3,
 					},
 					"v3/market/insurance": map[string]any{
-						"cost": 2 / 3,
+						"cost": float64(2) / 3,
 					},
 					"v3/market/riskLimit": map[string]any{
-						"cost": 2 / 3,
+						"cost": float64(2) / 3,
 					},
 					"v3/market/limitPrice": map[string]any{
-						"cost": 2 / 3,
+						"cost": float64(2) / 3,
 					},
 				},
 			},
@@ -624,7 +624,7 @@ func (this *Poloniex) Describe() any {
 					"marginMode":       false,
 					"limit":            100,
 					"daysBack":         nil,
-					"daysBackCanceled": 1 / 6,
+					"daysBackCanceled": float64(1) / 6,
 					"untilDays":        nil,
 					"trigger":          false,
 					"trailing":         false,
